Várias mudanças foram feitas. Vide README.md
Foram criados packages.
	Controller foi separado de Event
	Não consegui separar Treinador de FightControls

Várias linhas muito extensas e confusas foram melhor organizadas, de
forma que o código ficou mais claro.
Tabulação e espaçamento foram corrigidos

Reset() foi apagado.
